requirements.txt
asgiref==3.8.1
Django==5.0.6
django-jazzmin==3.0.0
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
